monitor/requirements.txt
google-genai>=0.6.0
watchdog>=5.0.0
PyMuPDF>=1.24.0
